neuralabs-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.4.2
psycopg2-binary==2.9.9
PyYAML==6.0.1
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
requests==2.31.0
aiohttp
web3==6.11.0
eth-account==0.10.0
eth-typing==4.0.0
hexbytes==0.3.1
cryptography>=3.4.8
websockets==12.0
cdp==1.0.0